requirements.txt
# requirements.txt
PyRPlidar
pyzmq
numpy
matplotlib
pySerial